Moved translations module and config
