--- src/config.ts
@@ -1,3 +1,0 @@
-export default {
-  backendHost: process.env.BACKEND_HOST || 'http://localhost:3000',
-};
