+++ src/config/config.json
@@ -1,0 +1,3 @@
+{
+  "backendHost": "http://localhost:3000"
+}
